deleted 2 files and small change
fixing bugs

rename pair left/right

fixed issues
