+++ src/com/amazon/tools/cli/ComparisonType.java
@@ -1,4 +1,4 @@
-package tools.cli;
+package com.amazon.tools.cli;
 
 public enum ComparisonType {
     BASIC,
