+++ src/com/amazon/tools/cli/IonJavaCli.java
@@ -1,69 +1,148 @@
-package tools.cli;
+package com.amazon.tools.cli;
 
 import com.amazon.ion.IonException;
-import com.amazon.ion.IonList;
 import com.amazon.ion.IonReader;
-import com.amazon.ion.IonSexp;
+import com.amazon.ion.IonSequence;
 import com.amazon.ion.IonStruct;
 import com.amazon.ion.IonSystem;
 import com.amazon.ion.IonTimestamp;
 import com.amazon.ion.IonType;
 import com.amazon.ion.IonValue;
 import com.amazon.ion.IonWriter;
+import com.amazon.ion.SymbolTable;
 import com.amazon.ion.SymbolToken;
 import com.amazon.ion.impl._Private_Utils;
 import com.amazon.ion.system.IonReaderBuilder;
 import com.amazon.ion.system.IonSystemBuilder;
+import com.amazon.ion.system.IonTextWriterBuilder;
 import com.amazon.ion.util.Equivalence;
-import com.sun.tools.javac.util.Pair;
 import org.kohsuke.args4j.Argument;
 import org.kohsuke.args4j.CmdLineException;
 import org.kohsuke.args4j.CmdLineParser;
 import org.kohsuke.args4j.Option;
-
-import tools.comparisonReport.ComparisonContext;
-import tools.comparisonReport.ComparisonResult;
-import tools.comparisonReport.ComparisonResultType;
-import tools.errorReport.ErrorDescription;
-import tools.errorReport.ErrorType;
-import tools.events.Event;
-import tools.events.EventType;
-import tools.events.ImportDescriptor;
+import com.amazon.tools.comparisonReport.ComparisonContext;
+import com.amazon.tools.comparisonReport.ComparisonResult;
+import com.amazon.tools.comparisonReport.ComparisonResultType;
+import com.amazon.tools.errorReport.ErrorDescription;
+import com.amazon.tools.errorReport.ErrorType;
+import com.amazon.tools.events.Event;
+import com.amazon.tools.events.EventType;
+import com.amazon.tools.events.ImportDescriptor;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.FilterOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.List;
 
-public class IonCompare {
+public class IonJavaCli {
     private static final int CONSOLE_WIDTH = 120; // Only used for formatting the USAGE message
     private static final int USAGE_ERROR_EXIT_CODE = 1;
+    private static final int IO_ERROR_EXIT_CODE = 2;
     private static final int BUFFER_SIZE = 128 * 1024;
     private static final String SYSTEM_OUT_DEFAULT_VALUE = "out";
     private static final String SYSTEM_ERR_DEFAULT_VALUE = "err";
+    private static final IonTextWriterBuilder ION_TEXT_WRITER_BUILDER = IonTextWriterBuilder.standard();
     private static final IonSystem ION_SYSTEM = IonSystemBuilder.standard().build();
     private static final String EMBEDDED_STREAM_ANNOTATION = "embedded_documents";
     private static final String EVENT_STREAM = "$ion_event_stream";
 
     public static void main(final String[] args) {
-        CompareArgs parsedArgs = new CompareArgs();
+        CommandArgs parsedArgs = new CommandArgs();
         CmdLineParser parser = new CmdLineParser(parsedArgs);
         parser.getProperties().withUsageWidth(CONSOLE_WIDTH);
-        ComparisonType comparisonType = parsedArgs.getComparisonType();
+        ComparisonType comparisonType = null;
         OutputFormat outputFormat = null;
+        CommandType commandType = null;
 
         try {
+            //parse
             parser.parseArgument(args);
+            validateArgs(parsedArgs);
+            //get value
             comparisonType = parsedArgs.getComparisonType();
             outputFormat = parsedArgs.getOutputFormat();
+            commandType = parsedArgs.getCommand();
         } catch (CmdLineException | IllegalArgumentException e) {
-            System.err.println(e.getMessage() + "\n");
+            printHelpText(e.getMessage(), parser, parsedArgs);
+            System.exit(USAGE_ERROR_EXIT_CODE);
+        }
+
+        ProcessContext processContext = commandType == CommandType.PROCESS ? new ProcessContext(null, -1,
+                null, null, null) : null;
+        try (
+                //Initialize output stream, never return null. (default value: STDOUT)
+                OutputStream outputStream = initOutputStream(parsedArgs, SYSTEM_OUT_DEFAULT_VALUE, processContext);
+                //Initialize error report, never return null. (default value: STDERR)
+                OutputStream errorReportOutputStream =
+                        initOutputStream(parsedArgs, SYSTEM_ERR_DEFAULT_VALUE, processContext);
+                IonWriter ionWriterForErrorReport = outputFormat.createIonWriter(errorReportOutputStream);
+                IonWriter ionWriterForOutput = outputFormat.createIonWriter(outputStream);
+        ) {
+            if (commandType == CommandType.COMPARE) {
+                compareFiles(ionWriterForOutput, ionWriterForErrorReport, parsedArgs, comparisonType);
+            } else if (commandType == CommandType.PROCESS) {
+                processContext.setIonWriter(ionWriterForOutput);
+                processFiles(ionWriterForErrorReport, parsedArgs, processContext);
+            }
+        } catch (IOException e) {
+            System.err.println("Failed to close OutputStream: " + e.getMessage());
+        } finally {
+            if (processContext != null) {
+                IonWriter ionWriter = processContext.getIonWriter();
+                if (ionWriter != null) {
+                    try {
+                        ionWriter.close();
+                    } catch (IOException e) {
+                        System.err.println(e.getMessage());
+                        System.exit(IO_ERROR_EXIT_CODE);
+                    }
+                }
+            }
+        }
+    }
+
+    private static void validateArgs(CommandArgs parsedArgs) throws CmdLineException {
+        parsedArgs.setCommand();
+        CommandType type = parsedArgs.getCommand();
+        switch (type) {
+            case PROCESS:
+                break;
+            case COMPARE:
+                if (parsedArgs.getOutputFormat() == OutputFormat.EVENTS) {
+                    throw new CmdLineException("COMPARE doesn't support option format \"-f events\"");
+                }
+                break;
+        }
+    }
+
+    private static void printHelpText(String msg, CmdLineParser parser, CommandArgs parsedArgs) {
+        CommandType commandType = parsedArgs.getCommand();
+
+        if (commandType == null) {
+            System.err.println(msg + "\n");
+            System.err.println("Usage : \n" +
+                    " [-h|--help] [-v|--version]\n" +
+                    "  -h, --help                print the help message and exit\n" +
+                    "  -v, --version             print the command's version number and exit");
+        } else if (commandType == CommandType.PROCESS) {
+            System.err.println(msg + "\n");
+            System.err.println("\"Process\" reads the input file(s) and re-write in the format specified by " +
+                    "--output.\n");
+            System.err.println("Usage:\n");
+            System.err.println("ion process [--output <file>] [--error-report <file>] [--output-format \n"
+                    + "(text | pretty | binary | events | none)] [--catalog <file>]... [--imports <file>]... \n"
+                    + "[--perf-report <file>] [--filter <filter> | --traverse <file>]  [-] [<input_file>]...\n");
+            System.err.println("Options:\n");
+            parser.printUsage(System.err);
+        } else if (commandType == CommandType.COMPARE) {
+            System.err.println(msg + "\n");
             System.err.println("\"Compare\" Compare all inputs (which may contain Ion streams and/or EventStreams) \n"
                     + "against all other inputs using the Ion data model's definition of equality. Write a \n"
                     + "ComparisonReport to the output.\n");
@@ -72,25 +151,164 @@ public class IonCompare {
                     + "pretty | binary | none)]  [--catalog <file>]... [--comparison-type (basic | equivs | \n"
                     + "non-equivs | equiv-timeline)] [-] [<input_file>]...\n");
             parser.printUsage(System.err);
-            System.exit(USAGE_ERROR_EXIT_CODE);
-        }
-
-        try (//Initialize output stream, never return null. (default value: STDOUT)
-             OutputStream outputStream = initOutputStream(parsedArgs, SYSTEM_OUT_DEFAULT_VALUE);
-             //Initialize error report, never return null. (default value: STDERR)
-             OutputStream errorReportOutputStream = initOutputStream(parsedArgs, SYSTEM_ERR_DEFAULT_VALUE);
-             IonWriter ionWriterForOutput = outputFormat.createIonWriter(outputStream);
-             IonWriter ionWriterForErrorReport = outputFormat.createIonWriter(errorReportOutputStream);
-        ) {
-            compareFiles(ionWriterForOutput, ionWriterForErrorReport, parsedArgs, comparisonType);
-        } catch (IOException e) {
-            System.err.println("Failed to close OutputStream: " + e.getMessage());
         }
     }
 
+    //
+    //
+    // functions for processing
+    //
+    //
+
+    private static void processFiles(IonWriter ionWriterForErrorReport,
+                                     CommandArgs args,
+                                     ProcessContext processContext) throws IOException {
+        if (args.getOutputFormat() == OutputFormat.EVENTS) {
+            processContext.getIonWriter().writeSymbol(EVENT_STREAM);
+        }
+
+        for (String path : args.getInputFiles()) {
+            try (
+                    InputStream inputStream = new BufferedInputStream(new FileInputStream(path));
+                    IonReader ionReader = IonReaderBuilder.standard().build(inputStream);
+            ) {
+                processContext.setFileName(path);
+
+                boolean isEventStream = isEventStream(ionReader);
+                List<Event> events = getEventStream(ionReader, isEventStream, CommandType.PROCESS);
+
+                processContext.setEventStream(events);
+                processContext.setEventIndex(0);
+
+                if (args.getOutputFormat() == OutputFormat.EVENTS) {
+                    processToEventStream(ionWriterForErrorReport, processContext);
+                } else {
+                    processToIonStream(processContext, args);
+                }
+                processContext.getIonWriter().finish();
+                ionWriterForErrorReport.finish();
+            } catch (IonException | NullPointerException e) {
+                new ErrorDescription(processContext.getState(), e.getMessage(), processContext.getFileName(),
+                        processContext.getEventIndex()).writeOutput(ionWriterForErrorReport);
+                System.exit(IO_ERROR_EXIT_CODE);
+            } catch (Exception e) {
+                new ErrorDescription(ErrorType.STATE, e.getMessage(), processContext.getFileName(),
+                        processContext.getEventIndex()).writeOutput(ionWriterForErrorReport);
+                System.exit(IO_ERROR_EXIT_CODE);
+            }
+        }
+    }
+
+    private static void processToEventStream(IonWriter ionWriterForErrorReport,
+                                             ProcessContext processContext) throws IOException {
+        for (Event event : processContext.getEventStream()) {
+            event.writeOutput(ionWriterForErrorReport, processContext);
+        }
+    }
+
+    private static void processToIonStream(ProcessContext processContext,
+                                           CommandArgs args) throws IOException {
+        List<Event> events = processContext.getEventStream();
+        int count = 0;
+        while (count != events.size()) {
+            //update eventIndex
+            Event event = events.get(count);
+            processContext.setEventIndex(processContext.getEventIndex() + 1);
+            processContext.setLastEventType(event.getEventType());
+            if (event.getEventType() == EventType.CONTAINER_START) {
+                if (isEmbeddedEvent(event)) {
+                    count = embeddedEventToIon(processContext, args, event.getIonType());
+                } else {
+                    IonType type = event.getIonType();
+                    setFieldName(event, processContext.getIonWriter());
+                    setAnnotations(event, processContext.getIonWriter());
+                    processContext.getIonWriter().stepIn(type);
+                }
+            } else if (event.getEventType().equals(EventType.CONTAINER_END)) {
+                processContext.getIonWriter().stepOut();
+            } else if (event.getEventType().equals(EventType.SCALAR)) {
+                writeIonByType(event, processContext.getIonWriter());
+            } else if (event.getEventType().equals(EventType.SYMBOL_TABLE)) {
+                handleSymbolTableEvent(processContext, event, args, false);
+            } else if (event.getEventType().equals(EventType.STREAM_END)) {
+                processContext.getIonWriter().finish();
+            }
+            count++;
+        }
+    }
+
+    private static int embeddedEventToIon(ProcessContext processContext,
+                                           CommandArgs args,
+                                           IonType ionType) throws IOException {
+        processContext.getIonWriter().addTypeAnnotation(EMBEDDED_STREAM_ANNOTATION);
+        processContext.getIonWriter().stepIn(ionType);
+        List<Event> events =  processContext.getEventStream();
+        int depth = 1;
+        boolean finish = false;
+        int count = 0;
+        while (++count < events.size()) {
+            StringBuilder out = new StringBuilder();
+            ProcessContext embeddedContext = new ProcessContext(null,0,null, null,
+                    ION_TEXT_WRITER_BUILDER.withImports(_Private_Utils.systemSymtab(1)).build(out));
+            embeddedContext.setEmbeddedOut(out);
+            try {
+                do {
+                    Event event = events.get(count);
+                    processContext.setEventIndex(processContext.getEventIndex() + 1);
+                    processContext.setLastEventType(event.getEventType());
+                    if (event.getEventType() == EventType.STREAM_END) {
+                        break;
+                    } else if (event.getEventType() == EventType.SCALAR) {
+                        writeIonByType(event, embeddedContext.getIonWriter());
+                    } else if (event.getEventType() == EventType.CONTAINER_START) {
+                        depth++;
+                        embeddedContext.getIonWriter().stepIn(event.getIonType());
+                    } else if (event.getEventType() == EventType.CONTAINER_END) {
+                        depth--;
+                        if (depth == 0) {
+                            if (event.getIonType() == IonType.SEXP || event.getIonType() == IonType.LIST) {
+                                finish = true;
+                                break;
+                            } else {
+                                throw new IonException("invalid CONTAINER_END");
+                            }
+                        }
+                        embeddedContext.getIonWriter().stepOut();
+                    } else if (event.getEventType() == EventType.SYMBOL_TABLE) {
+                        handleSymbolTableEvent(embeddedContext, event, args, true);
+                    }
+                } while (++count < events.size());
+
+                if (!finish) {
+                    embeddedContext.getIonWriter().finish();
+                    processContext.getIonWriter().writeString(out.toString());
+                }
+            } finally {
+                IonWriter ionWriter = embeddedContext.getIonWriter();
+                if (ionWriter != null) {
+                    try {
+                        ionWriter.close();
+                    } catch (IOException e) {
+                        System.err.println(e.getMessage());
+                        System.exit(IO_ERROR_EXIT_CODE);
+                    }
+                }
+            }
+            if (finish) { break; }
+        }
+        processContext.getIonWriter().stepOut();
+        return count;
+    }
+
+    //
+    //
+    //  functions for comparing
+    //
+    //
+
     private static void compareFiles(IonWriter ionWriterForOutput,
                                      IonWriter ionWriterForErrorReport,
-                                     CompareArgs args,
+                                     CommandArgs args,
                                      ComparisonType comparisonType) throws IOException {
         List<String> files = args.getInputFiles();
         CompareContext compareContext = new CompareContext(null, null);
@@ -104,13 +322,13 @@ public class IonCompare {
                      IonReader ionReaderSecond = IonReaderBuilder.standard().build(inputSecond);
                 ) {
                     if (comparisonType == ComparisonType.BASIC) {
-                        if (path.equals(compareToPath)) continue;
+                        if (path.equals(compareToPath)) { continue; }
                     }
 
                     boolean isFirstEventStream = isEventStream(ionReaderFirst);
                     boolean isSecondEventStream = isEventStream(ionReaderSecond);
-                    ArrayList<Event> eventsFirst = getEventStream(ionReaderFirst, isFirstEventStream);
-                    ArrayList<Event> eventsSecond = getEventStream(ionReaderSecond, isSecondEventStream);
+                    List<Event> eventsFirst = getEventStream(ionReaderFirst, isFirstEventStream, CommandType.COMPARE);
+                    List<Event> eventsSecond = getEventStream(ionReaderSecond, isSecondEventStream, CommandType.COMPARE);
                     compareContext.setEventStreamFirst(eventsFirst);
                     compareContext.setEventStreamSecond(eventsSecond);
 
@@ -131,24 +349,17 @@ public class IonCompare {
                         }
                     }
                 } catch (IonException | NullPointerException e) {
-                    new ErrorDescription(ErrorType.STATE, e.getMessage(), path+';'+compareToPath,
+                    new ErrorDescription(ErrorType.STATE, e.getMessage(), path + ';' + compareToPath,
                             -1).writeOutput(ionWriterForErrorReport);
                 }
             }
         }
     }
 
-    public static boolean isEventStream(IonReader ionReader) {
-        return ionReader.next() != null
-                && ionReader.getType() == IonType.SYMBOL
-                && EVENT_STREAM.equals(ionReader.symbolValue().getText());
-    }
-
-
     private static boolean compare(CompareContext compareContext,
                                    int startI, int endI, int startJ, int endJ) throws IOException {
-        ArrayList<Event> eventsFirst = (ArrayList<Event>) compareContext.getEventStreamFirst();
-        ArrayList<Event> eventsSecond = (ArrayList<Event>) compareContext.getEventStreamSecond();
+        List<Event> eventsFirst = compareContext.getEventStreamFirst();
+        List<Event> eventsSecond = compareContext.getEventStreamSecond();
         int i = startI;
         int j = startJ;
         while (i <= endI && j <= endJ && i < eventsFirst.size() && j < eventsSecond.size()) {
@@ -161,19 +372,19 @@ public class IonCompare {
             EventType eventTypeFirst = eventFirst.getEventType();
             EventType eventTypeSecond = eventSecond.getEventType();
             if (eventTypeFirst != eventTypeSecond) {
-                setReportInfo(i, j,"Did't match event_type", compareContext);
+                setReportInfo(i, j,"Didn't match event_type", compareContext);
                 return false;
             } else if (eventFirst.getDepth() != eventSecond.getDepth()) {
-                setReportInfo(i, j,"Did't match depth", compareContext);
+                setReportInfo(i, j,"Didn't match depth", compareContext);
                 return false;
             } else if (eventFirst.getIonType() != eventSecond.getIonType()) {
-                setReportInfo(i, j,"Did't match ion_type", compareContext);
+                setReportInfo(i, j,"Didn't match ion_type", compareContext);
                 return false;
             } else if (!isSameSymbolToken(fieldNameFirst, fieldNameSecond)) {
-                setReportInfo(i, j,"Did't match field_name", compareContext);
+                setReportInfo(i, j,"Didn't match field_name", compareContext);
                 return false;
             } else if (!isSameSymbolTokenArray(annotationFirst, annotationSecond)) {
-                setReportInfo(i, j,"Did't match annotation", compareContext);
+                setReportInfo(i, j,"Didn't match annotation", compareContext);
                 return false;
             }
 
@@ -226,7 +437,6 @@ public class IonCompare {
         if (initEvent.getEventType() != EventType.CONTAINER_START || initEvent.getIonType() != IonType.STRUCT) {
             return null;
         }
-
         IonStruct ionStruct = ION_SYSTEM.newEmptyStruct();
         while (containerContext.increaseIndex() < end) {
             Event event = isFirst ?
@@ -237,7 +447,7 @@ public class IonCompare {
                 switch (event.getIonType()) {
                     case LIST:
                         ionStruct.add(event.getFieldName(),
-                                parseList(containerContext, compareContext, end, isFirst));
+                                parseSequence(containerContext, compareContext, end, isFirst, ION_SYSTEM.newEmptyList()));
                         break;
                     case STRUCT:
                         ionStruct.add(event.getFieldName(),
@@ -245,7 +455,7 @@ public class IonCompare {
                         break;
                     case SEXP:
                         ionStruct.add(event.getFieldName(),
-                                parseSexp(containerContext, compareContext, end, isFirst));
+                                parseSequence(containerContext, compareContext, end, isFirst, ION_SYSTEM.newEmptySexp()));
                         break;
                 }
             } else if (eventType == EventType.CONTAINER_END) {
@@ -253,20 +463,19 @@ public class IonCompare {
             } else if (eventType == EventType.STREAM_END) {
                 throw new IonException("Invalid struct: eventStream ends without CONTAINER_END");
             } else {
-                IonValue value = event.getValue();
-                value.removeFromContainer();
-                value.setTypeAnnotationSymbols(event.getAnnotations());
-                ionStruct.add(event.getFieldName(), event.getValue());
+                IonValue cloneValue = event.getValue().clone();
+                cloneValue.setTypeAnnotationSymbols(event.getAnnotations());
+                ionStruct.add(event.getFieldName(), cloneValue);
             }
         }
         return ionStruct;
     }
 
-    private static IonList parseList(ContainerContext containerContext,
-                                               CompareContext compareContext,
-                                               int end,
-                                               boolean isFirst) {
-        IonList ionList = ION_SYSTEM.newEmptyList();
+    private static IonSequence parseSequence(ContainerContext containerContext,
+                                             CompareContext compareContext,
+                                             int end,
+                                             boolean isFirst,
+                                             IonSequence ionSequence) {
         while (containerContext.increaseIndex() < end) {
             Event event = isFirst ?
                     compareContext.getEventStreamFirst().get(containerContext.getIndex()) :
@@ -275,82 +484,33 @@ public class IonCompare {
             if (eventType == EventType.CONTAINER_START) {
                 switch (event.getIonType()) {
                     case LIST:
-                        ionList.add(parseList(containerContext, compareContext, end, isFirst));
-                        break;
-                    case STRUCT:
-                        ionList.add(parseStruct(containerContext, compareContext, end, isFirst));
+                        ionSequence.add(parseSequence(containerContext, compareContext, end, isFirst, ION_SYSTEM.newEmptyList()));
                         break;
                     case SEXP:
-                        ionList.add(parseSexp(containerContext, compareContext, end, isFirst));
-                }
-            } else if (eventType == EventType.CONTAINER_END) {
-                break;
-            } else if (eventType == EventType.STREAM_END) {
-                throw new IonException("Invalid list: eventStream ends without CONTAINER_END");
-            } else {
-                IonValue value = event.getValue();
-                value.removeFromContainer();
-                value.setTypeAnnotationSymbols(event.getAnnotations());
-                ionList.add(event.getValue());
-            }
-        }
-        return ionList;
-    }
-
-    private static IonSexp parseSexp(ContainerContext containerContext,
-                                          CompareContext compareContext,
-                                          int end,
-                                          boolean isFirst) {
-        IonSexp ionSexp = ION_SYSTEM.newEmptySexp();
-        while (containerContext.increaseIndex() < end) {
-            Event event = isFirst ?
-                    compareContext.getEventStreamFirst().get(containerContext.getIndex()) :
-                    compareContext.getEventStreamSecond().get(containerContext.getIndex());
-            EventType eventType = event.getEventType();
-            if (eventType == EventType.CONTAINER_START) {
-                switch (event.getIonType()) {
-                    case LIST:
-                        ionSexp.add(parseList(containerContext, compareContext, end, isFirst));
+                        ionSequence.add(parseSequence(containerContext, compareContext, end, isFirst, ION_SYSTEM.newEmptySexp()));
                         break;
                     case STRUCT:
-                        ionSexp.add(parseStruct(containerContext, compareContext, end, isFirst));
-                        break;
-                    case SEXP:
-                        ionSexp.add(parseSexp(containerContext, compareContext, end, isFirst));
+                        ionSequence.add(parseStruct(containerContext, compareContext, end, isFirst));
                         break;
                 }
             } else if (eventType == EventType.CONTAINER_END) {
                 break;
             } else if (eventType == EventType.STREAM_END) {
-                throw new IonException("Invalid sexp: eventStream ends without CONTAINER_END");
+                throw new IonException("Invalid ionSequence: eventStream ends without CONTAINER_END");
             } else {
-                IonValue value = event.getValue();
-                value.removeFromContainer();
-                value.setTypeAnnotationSymbols(event.getAnnotations());
-                ionSexp.add(event.getValue());
+                IonValue cloneValue = event.getValue().clone();
+                cloneValue.setTypeAnnotationSymbols(event.getAnnotations());
+                ionSequence.add(cloneValue);
             }
         }
-        return ionSexp;
-    }
-
-    private static void writeReport(CompareContext compareContext,
-                                    IonWriter ionWriter,
-                                    ComparisonResultType type) throws IOException {
-        new ComparisonResult(type,
-                new ComparisonContext(compareContext.getFile(),
-                        compareContext.getEventStreamFirst().get(compareContext.getFileEventIndex()),
-                        compareContext.getFileEventIndex() + 1),
-                new ComparisonContext(compareContext.getCompareToFile(),
-                        compareContext.getEventStreamSecond().get(compareContext.getCompareToFileEventIndex()),
-                        compareContext.getCompareToFileEventIndex() + 1),
-                compareContext.getMessage()).writeOutput(ionWriter);
+        return ionSequence;
     }
 
     private static boolean compareEquivs(CompareContext compareContext) throws IOException {
         int i = 0;
         int j = 0;
-        ArrayList<Event> eventStreamFirst = (ArrayList<Event>) compareContext.getEventStreamFirst();
-        ArrayList<Event> eventStreamSecond = (ArrayList<Event>) compareContext.getEventStreamSecond();
+        List<Event> eventStreamFirst = compareContext.getEventStreamFirst();
+        List<Event> eventStreamSecond = compareContext.getEventStreamSecond();
         ComparisonType type = compareContext.getType();
 
         while (i < eventStreamFirst.size() && j < eventStreamSecond.size()) {
@@ -372,28 +532,28 @@ public class IonCompare {
                 throw new IonException("Embedded streams set expected.");
             }
 
-            ArrayList<Pair<Integer, Integer>> pairsFirst;
-            ArrayList<Pair<Integer, Integer>> pairsSecond;
+            List<Pair> pairsFirst;
+            List<Pair> pairsSecond;
 
             if (isEmbeddedEvent(eventFirst) && isEmbeddedEvent(eventSecond)) {
-                pairsFirst = parseEmbeddedStream(eventStreamFirst, i);
-                pairsSecond = parseEmbeddedStream(eventStreamSecond, j);
+                pairsFirst = locateEmbeddedStreamBoundaries(eventStreamFirst, i);
+                pairsSecond = locateEmbeddedStreamBoundaries(eventStreamSecond, j);
             } else {
-                pairsFirst = parseContainer(eventStreamFirst, i);
-                pairsSecond = parseContainer(eventStreamSecond, j);
+                pairsFirst = locateContainerBoundaries(eventStreamFirst, i);
+                pairsSecond = locateContainerBoundaries(eventStreamSecond, j);
             }
-            i = pairsFirst.size() == 0 ? i + 1 : pairsFirst.get(pairsFirst.size() - 1).snd + 1;
-            j = pairsSecond.size() == 0 ? j + 1 : pairsSecond.get(pairsSecond.size() - 1).snd + 1;
+            i = pairsFirst.size() == 0 ? i + 1 : pairsFirst.get(pairsFirst.size() - 1).right + 1;
+            j = pairsSecond.size() == 0 ? j + 1 : pairsSecond.get(pairsSecond.size() - 1).right + 1;
 
             for (int m = 0; m < pairsFirst.size(); m++) {
                 for (int n = 0; n < pairsSecond.size(); n++) {
                     if (m == n) continue;
-                    Pair<Integer, Integer> pairFirst = pairsFirst.get(m);
-                    Pair<Integer, Integer> pairSecond = pairsSecond.get(n);
-                    if (compare(compareContext, pairFirst.fst, pairFirst.snd, pairSecond.fst, pairSecond.snd) ^
+                    Pair pairFirst = pairsFirst.get(m);
+                    Pair pairSecond = pairsSecond.get(n);
+                    if (compare(compareContext, pairFirst.left, pairFirst.right, pairSecond.left, pairSecond.right) ^
                             (type == ComparisonType.EQUIVS || type == ComparisonType.EQUIVS_TIMELINE)) {
                         if (type == ComparisonType.NON_EQUIVS) {
-                            setReportInfo(pairFirst.fst, pairSecond.fst,
+                            setReportInfo(pairFirst.left, pairSecond.left,
                                     "Equivalent values in a non-equivs set.", compareContext);
                         }
                         return type == ComparisonType.NON_EQUIVS;
@@ -407,23 +567,24 @@ public class IonCompare {
     }
 
     /**
-     *  This function will parse a embedded stream into pairs<int,int> that stores index range for
-     *  each string it includes.
+     *  This function parses an embedded stream into pairs<int,int> that stores index range for
+     *  each string the embedded stream includes.
      *
      *  Input value 'start' must be index of a CONTAINER_START event.
      *
      *  return List is never null. It will return a list with size zero if the embedded stream is empty.
      */
-    private static ArrayList<Pair<Integer, Integer>> parseEmbeddedStream(ArrayList<Event> events, int start) {
-        ArrayList<Pair<Integer, Integer>> parsedEvents = new ArrayList<>(0);
-        int count = start + 1;
+    private static List<Pair> locateEmbeddedStreamBoundaries(List<Event> events, int start) {
+        List<Pair> parsedEvents = new ArrayList<>(0);
+        int left = start + 1;
+        int right = start;
         int depth = 1;
-        while (++start < events.size()) {
-            EventType eventType = events.get(start).getEventType();
+        while (++right < events.size()) {
+            EventType eventType = events.get(right).getEventType();
 
             if (eventType == EventType.STREAM_END) {
-                parsedEvents.add(new Pair<>(count, start));
-                count = start + 1;
+                parsedEvents.add(new Pair(left, right));
+                left = right + 1;
             } else if (eventType == EventType.CONTAINER_END) {
                 if (--depth == 0) break;
             } else if (eventType == EventType.CONTAINER_START) {
@@ -434,8 +595,8 @@ public class IonCompare {
     }
 
     /**
-     *  This function will parse a container's values into pairs<int,int> that stores its index range.
-     *  1. For SCALAR event with index i, the pair will be (i, i)
+     *  This function parses a container's values into pairs<int,int> that stores its index range.
+     *  1. For SCALAR event with index i, the pair will be (i, i).
      *  2. For CONTAINER event, the pair will be (start, end) where start is the index of the CONTAINER_START event
      *  and end is the index of the CONTAINER_END event.
      *
@@ -443,8 +604,8 @@ public class IonCompare {
      *
      *  return List is never null. It will return a list with size zero if the container is empty.
      */
-    private static ArrayList<Pair<Integer, Integer>> parseContainer(ArrayList<Event> events, int start) {
-        ArrayList<Pair<Integer, Integer>> parsedEvents = new ArrayList<>(0);
+    private static List<Pair> locateContainerBoundaries(List<Event> events, int start) {
+        List<Pair> parsedEvents = new ArrayList<>(0);
         while (++start < events.size()) {
             EventType eventType = events.get(start).getEventType();
 
@@ -453,23 +614,25 @@ public class IonCompare {
             } else if (eventType == EventType.CONTAINER_END) {
                 break;
             } else if (eventType == EventType.SCALAR) {
-                parsedEvents.add(new Pair<>(start, start));
+                parsedEvents.add(new Pair(start, start));
             } else if (eventType == EventType.CONTAINER_START) {
                 int startCount = start;
                 int endCount = findNextContainer(events, start);
                 start = endCount;
-                parsedEvents.add(new Pair<>(startCount, endCount));
+                parsedEvents.add(new Pair(startCount, endCount));
             }
         }
         return parsedEvents;
     }
 
     /**
-     *  This function will move the count i from the CONTAINER_START event to its corresponding CONTAINER_END event
+     *  This function will move the count i from the CONTAINER_START event to its corresponding CONTAINER_END event.
      *
-     *  return int is the index of its corresponding CONTAINER_END event.
+     *  return the index of its corresponding CONTAINER_END event.
+     *
+     *  throw IonException if CONTAINER_END event doesn't exist.
      */
-    private static int findNextContainer(ArrayList<Event> eventStream, int i) {
+    private static int findNextContainer(List<Event> eventStream, int i) {
         if (!IonType.isContainer(eventStream.get(i).getIonType())) return i;
         int depth = 1;
         while (i++ < eventStream.size()) {
@@ -486,9 +649,63 @@ public class IonCompare {
         return i;
     }
 
+    //
+    //
+    // helper functions
+    //
+    //
+
+    private static void handleSymbolTableEvent(ProcessContext processContext,
+                                               Event event,
+                                               CommandArgs args,
+                                               boolean isEmbedded) throws IOException {
+        processContext.getIonWriter().close();
+        ImportDescriptor[] imports = event.getImports();
+        SymbolTable[] symbolTables = new SymbolTable[imports.length];
+        for (int i = 0; i < imports.length; i++) {
+            SymbolTable symbolTable = ION_SYSTEM.newSharedSymbolTable(
+                    imports[i].getImportName(), imports[i].getVersion(), null);
+            symbolTables[i] = symbolTable;
+        }
+        if (!isEmbedded) {
+            processContext.setIonWriter(args.getOutputFormat().
+                    createIonWriterWithImports(new FileOutputStream(processContext.getFile(), true), symbolTables));
+        } else {
+            processContext.getEmbeddedOut().append(" ");
+            processContext.setIonWriter(ION_TEXT_WRITER_BUILDER.withImports(symbolTables)
+                    .build(processContext.getEmbeddedOut()));
+        }
+    }
+
+    private static void writeIonByType(Event event, IonWriter ionWriter) {
+        setFieldName(event, ionWriter);
+        IonValue value = event.getValue();
+        value.setTypeAnnotationSymbols(event.getAnnotations());
+        value.writeTo(ionWriter);
+    }
+
+    private static void setFieldName(Event event, IonWriter ionWriter) {
+        SymbolToken field = event.getFieldName();
+        if (field != null) {
+            if (!ionWriter.isInStruct()) throw new IonException("invalid field_name inside STRUCT");
+            ionWriter.setFieldNameSymbol(field);
+        } else if (ionWriter.isInStruct()) {
+            SymbolToken symbolToken = _Private_Utils.newSymbolToken((String) null, 0);
+            ionWriter.setFieldNameSymbol(symbolToken);
+        }
+    }
+
+    private static void setAnnotations(Event event, IonWriter ionWriter) {
+        SymbolToken[] annotations = event.getAnnotations();
+        if (annotations != null && annotations.length != 0) {
+            ionWriter.setTypeAnnotationSymbols(annotations);
+        }
+    }
+
     private static boolean isSameSymbolToken(SymbolToken symbolTokenX, SymbolToken symbolTokenY) {
-        if (symbolTokenX == null && symbolTokenY == null) return true;
-        else if (symbolTokenX != null && symbolTokenY != null) {
+        if (symbolTokenX == null && symbolTokenY == null) {
+            return true;
+        } else if (symbolTokenX != null && symbolTokenY != null) {
             if (symbolTokenX.getText() != null && symbolTokenY.getText() != null) {
                 return symbolTokenX.getText().equals(symbolTokenY.getText());
             } else {
@@ -498,8 +715,9 @@ public class IonCompare {
     }
 
     private static boolean isSameSymbolTokenArray(SymbolToken[] symbolTokenArrayX, SymbolToken[] symbolTokenArrayY) {
-        if (symbolTokenArrayX == null && symbolTokenArrayY == null) return true;
-        else if (symbolTokenArrayX != null && symbolTokenArrayY != null) {
+        if (symbolTokenArrayX == null && symbolTokenArrayY == null) {
+            return true;
+        } else if (symbolTokenArrayX != null && symbolTokenArrayY != null) {
             if (symbolTokenArrayX.length == symbolTokenArrayY.length) {
                 for (int i = 0; i < symbolTokenArrayX.length; i++) {
                     if (!isSameSymbolToken(symbolTokenArrayX[i], symbolTokenArrayY[i])) return false;
@@ -530,13 +748,40 @@ public class IonCompare {
                 && annotations[0].getText().equals(EMBEDDED_STREAM_ANNOTATION);
     }
 
-    private static ArrayList<Event> getEventStream(IonReader ionReader, boolean isEventStream) throws IOException {
-        ArrayList<Event> events = new ArrayList<>();
+    public static boolean isEventStream(IonReader ionReader) {
+        return ionReader.next() != null
+                && ionReader.getType() == IonType.SYMBOL
+                && EVENT_STREAM.equals(ionReader.symbolValue().getText());
+    }
+
+    private static void setReportInfo(int i, int j, String message, CompareContext compareContext) {
+        if (i != -1) compareContext.setFileEventIndex(i);
+        if (j != -1) compareContext.setCompareToFileEventIndex(j);
+        if (message != null) compareContext.setMessage(message);
+    }
+
+    private static void writeReport(CompareContext compareContext,
+                                    IonWriter ionWriter,
+                                    ComparisonResultType type) throws IOException {
+        new ComparisonResult(type,
+                new ComparisonContext(compareContext.getFile(),
+                        compareContext.getEventStreamFirst().get(compareContext.getFileEventIndex()),
+                        compareContext.getFileEventIndex()),
+                new ComparisonContext(compareContext.getCompareToFile(),
+                        compareContext.getEventStreamSecond().get(compareContext.getCompareToFileEventIndex()),
+                        compareContext.getCompareToFileEventIndex()),
+                compareContext.getMessage()).writeOutput(ionWriter);
+    }
+
+    private static List<Event> getEventStream(IonReader ionReader,
+                                              boolean isEventStream,
+                                              CommandType commandType) throws IOException {
+        List<Event> events = new ArrayList<>();
 
         if (isEventStream) {
             while (ionReader.next() != null) {
                 Event event = eventStreamToEvent(ionReader);
-                if (event.getEventType() == EventType.SYMBOL_TABLE) {
+                if (event.getEventType() == EventType.SYMBOL_TABLE && commandType == CommandType.COMPARE) {
                     continue;
                 }
                 events.add(event);
@@ -550,7 +795,7 @@ public class IonCompare {
         return events;
     }
 
-    private static void validateEventStream(ArrayList<Event> events) {
+    private static void validateEventStream(List<Event> events) {
         if (events.get(events.size() - 1).getEventType() != EventType.STREAM_END) {
             throw new IonException("Invalid event stream: event stream end without STREAM_END event");
         }
@@ -581,9 +826,9 @@ public class IonCompare {
                         if (aEventType == EventType.CONTAINER_START) {
                             depth++;
                         } else if (aEventType == EventType.CONTAINER_END) {
-                           if (--depth == 0) {
-                               throw new IonException("Invalid EventStream: end without STREAM_END event");
-                           }
+                            if (--depth == 0) {
+                                throw new IonException("Invalid EventStream: end without STREAM_END event");
+                            }
                         } else if (aEventType == EventType.STREAM_END) {
                             if (depth != 1) {
                                 throw new IonException("Invalid EventStream: " +
@@ -604,11 +849,19 @@ public class IonCompare {
         }
     }
 
-    private static ArrayList<Event> ionStreamToEventStream(IonReader ionReader) throws IOException {
-        ArrayList<Event> events = new ArrayList<>();
+    private static List<Event> ionStreamToEventStream(IonReader ionReader) throws IOException {
+        List<Event> events = new ArrayList<>();
         if (ionReader.getType() == null ) return events;
+        SymbolTable curTable = ionReader.getSymbolTable();
 
         do {
+            if (!isSameSymbolTable(ionReader.getSymbolTable(), curTable)) {
+                curTable = ionReader.getSymbolTable();
+                ImportDescriptor[] imports = symbolTableToImports(curTable.getImportedTables());
+                events.add(new Event(EventType.SYMBOL_TABLE, null, null, null,
+                        null, imports, 0));
+            }
+
             if (isEmbeddedStream(ionReader)) {
                 //get current Ion type and depth
                 IonType curType = ionReader.getType();
@@ -625,7 +878,7 @@ public class IonCompare {
                     String stream = ionReader.stringValue();
                     try (IonReader tempIonReader = IonReaderBuilder.standard().build(stream)) {
                         while (tempIonReader.next() != null) {
-                            ArrayList<Event> append = ionStreamToEventStream(tempIonReader);
+                            List<Event> append = ionStreamToEventStream(tempIonReader);
                             events.addAll(append);
                         }
                     }
@@ -647,7 +900,7 @@ public class IonCompare {
 
                 //recursive call
                 ionReader.next();
-                ArrayList<Event> append = ionStreamToEventStream(ionReader);
+                List<Event> append = ionStreamToEventStream(ionReader);
                 events.addAll(append);
                 //write a Container_End event and step out
                 events.add(new Event(EventType.CONTAINER_END, curType, null, null,
@@ -661,15 +914,46 @@ public class IonCompare {
         return events;
     }
 
-    private static void setReportInfo(int i, int j, String message, CompareContext compareContext) {
-        if (i != -1) compareContext.setFileEventIndex(i);
-        if (j != -1) compareContext.setCompareToFileEventIndex(j);
-        if (message != null) compareContext.setMessage(message);
+    private static ImportDescriptor[] symbolTableToImports(SymbolTable[] tables) {
+        if (tables == null || tables.length == 0) return null;
+        int size = tables.length;
+
+        ImportDescriptor[] imports = new ImportDescriptor[size];
+        for (int i = 0; i < size; i++) {
+            ImportDescriptor table = new ImportDescriptor(tables[i]);
+            imports[i] = table;
+        }
+        return imports;
     }
 
-    /**
-     * this function is only used for CONTAINER_START and SCALAR eventType since other types don't need initial data.
-     */
+    private static boolean isSameSymbolTable(SymbolTable newTable, SymbolTable curTable) {
+        if (newTable == null && curTable == null) { return true; }
+        else if (newTable != null && curTable == null) { return false; }
+        else if (newTable == null) { return false; }
+
+        if (newTable.isLocalTable() && newTable.getImportedTables().length == 0) {
+            return true;
+        } else if (newTable.isSystemTable() && curTable.isSystemTable()) {
+            return newTable.getVersion() == curTable.getVersion();
+        } else if (newTable.isSharedTable() && curTable.isSharedTable()) {
+            return (newTable.getName().equals(curTable.getName())) & (newTable.getVersion() == curTable.getVersion());
+        } else if (newTable.isLocalTable() && curTable.isLocalTable()) {
+            SymbolTable[] xTable = newTable.getImportedTables();
+            SymbolTable[] yTable = curTable.getImportedTables();
+            //compare imports
+            if (xTable.length == yTable.length) {
+                for (int i = 0; i < xTable.length; i++) {
+                    if (!isSameSymbolTable(xTable[i], yTable[i]))  return false;
+                }
+            } else {
+                return false;
+            }
+            return true;
+        } else {
+            return false;
+        }
+    }
+
     private static Event ionStreamToEvent(IonReader ionReader) throws IllegalStateException {
         if (ionReader.getType() == null) throw new IllegalStateException("Can't convert ionReader null type to Event");
         IonType ionType = ionReader.getType();
@@ -840,6 +1124,7 @@ public class IonCompare {
                 } else if (fieldName != null && ionType != null) {
                     throw new IonException("Invalid SYMBOL_TABLE: unnecessary fields");
                 }
+                break;
             case CONTAINER_END:
                 if (depth == -1 || ionType == null) {
                     throw new IonException("Invalid CONTAINER_END: missing depth");
@@ -849,6 +1134,7 @@ public class IonCompare {
                 } else if (fieldName != null && imports != null) {
                     throw new IonException("Invalid CONTAINER_END: unnecessary fields");
                 }
+                break;
             case STREAM_END:
                 if (depth == -1) {
                     throw new IonException("Invalid STREAM_END: missing depth");
@@ -904,23 +1190,31 @@ public class IonCompare {
      *
      *  this function never return null
      */
-    private static BufferedOutputStream initOutputStream(CompareArgs args, String defaultValue) throws IOException {
+    private static BufferedOutputStream initOutputStream(CommandArgs args,
+                                                         String defaultValue,
+                                                         ProcessContext processContext) throws IOException {
         BufferedOutputStream outputStream = null;
         switch (defaultValue) {
             case SYSTEM_OUT_DEFAULT_VALUE:
                 String outputFile = args.getOutputFile();
                 if (outputFile != null && outputFile.length() != 0) {
                     File myFile = new File(outputFile);
+                    if (args.getCommand() == CommandType.PROCESS && processContext != null) {
+                        processContext.setFile(myFile);
+                    }
                     FileOutputStream out = new FileOutputStream(myFile);
                     outputStream = new BufferedOutputStream(out, BUFFER_SIZE);
                 } else {
-                    outputStream = new BufferedOutputStream(System.out, BUFFER_SIZE); }
+                    OutputStream out = new FilterOutputStream(System.out);
+                    out.close();
+                    outputStream = new BufferedOutputStream(System.out, BUFFER_SIZE);
+                }
                 break;
             case SYSTEM_ERR_DEFAULT_VALUE:
                 String errorReport = args.getErrorReport();
                 if (errorReport != null && errorReport.length() != 0) {
                     File myFile = new File(errorReport);
-                    FileOutputStream out = new FileOutputStream(myFile);
+                    FileOutputStream out = new FileOutputStream(myFile, true);
                     outputStream = new BufferedOutputStream(out, BUFFER_SIZE);
                 } else {
                     outputStream = new BufferedOutputStream(System.err, BUFFER_SIZE);
@@ -928,6 +1222,16 @@ public class IonCompare {
                 break;
         }
         return outputStream;
+    }
+
+    static class Pair {
+        final int left;
+        final int right;
+
+        Pair(int left, int right) {
+            this.left = left;
+            this.right = right;
+        }
     }
 
     static class ContainerContext {
@@ -951,7 +1255,7 @@ public class IonCompare {
         }
     }
 
-    static class CompareArgs {
+    static class CommandArgs {
         private static final String DEFAULT_FORMAT_VALUE = OutputFormat.PRETTY.toString();
         private static final String DEFAULT_COMPARISON_TYPE = ComparisonType.BASIC.toString();
 
@@ -991,12 +1295,25 @@ public class IonCompare {
                         + "equivalent by the Ion data model.")
         private String comparisonType = DEFAULT_COMPARISON_TYPE;
 
+        @Argument(required = true)
+        private List<String> inputs;
 
-       @Argument(metaVar = "FILES", required = true)
-        private List<String> inputFiles;
+        private CommandType commandType = null;
+
+        public String getOutputFormatName() {
+            return outputFormatName;
+        }
 
         public ComparisonType getComparisonType() throws IllegalArgumentException {
             return ComparisonType.valueOf(comparisonType.replace('-','_').toUpperCase());
+        }
+
+        public CommandType getCommand() {
+            return this.commandType;
+        }
+
+        public void setCommand() {
+            this.commandType = CommandType.valueOf(inputs.get(0).toUpperCase());
         }
 
         public OutputFormat getOutputFormat() throws IllegalArgumentException {
@@ -1004,7 +1321,9 @@ public class IonCompare {
         }
 
         public List<String> getInputFiles() {
-            return inputFiles;
+            int length = this.inputs.size();
+            List<String> files = this.inputs.subList(1, length);
+            return files;
         }
         public String getOutputFile() { return outputFile; }
         public String getErrorReport() { return errorReport; }
